feat: 用户 ID 改用雪花算法，新增随机密码生成功能
- User.ID 从自增 uint 改为雪花 ID (int64) 作为主键
- 前端 ID 以字符串传输，避免 JavaScript 大数精度丢失
- 分享提取码和用户密码输入框新增随机生成按钮
- 修复新建用户后提示"用户不存在"的问题

+++ internal/handler/setting_test.go
@@ -68,7 +68,7 @@ func setupSettingHandler(t *testing.T) (*gin.Engine, *service.JWTSecretStore, *m
 	return r, store, admin, user
 }
 
-func authHeader(t *testing.T, userID uint, secret string) string {
+func authHeader(t *testing.T, userID int64, secret string) string {
 	t.Helper()
 	tok, err := middleware.GenerateToken(userID, secret)
 	if err != nil {
